+++ orch.gemspec
@@ -23,5 +23,4 @@ Gem::Specification.new do |spec|
 
   spec.add_dependency 'thor', '~> 0.18'
   spec.add_dependency 'hashie'
-
 end
